complete MapxRawVersioned
