+++ Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 description = "Versioned Stateful DataBase, mainly used in blockchain scene."
 homepage = "https://github.com/ccmlm/vsdb"
 repository = "https://github.com/ccmlm/vsdb"
-keywords = ["version", "blockchain", "database"]
+keywords = ["vcs", "version management", "blockchain", "database"]
 categories = ["database"]
 license = "MIT"
